add part javadock

+++ src/rooppin/video/rental/ApplicationWindow.java
@@ -5,13 +5,21 @@ import java.net.MalformedURLException;
 import java.util.*;
 import javax.swing.*;
 
-@SuppressWarnings("serial")
+/**
+ * A logical continuation of the SearchAppFrame class.
+ * Implements all the functionality of links
+ * 
+ * @author _GoLeM_
+ *
+ */
 public class ApplicationWindow extends SearchAppFrame
 {
 	User user;
-	
 	protected MenuPanel menu;
 	
+	/**
+	 * Constructor
+	 */
 	public ApplicationWindow()
 	{
 		user = new User(this);
@@ -21,11 +29,23 @@ public class ApplicationWindow extends SearchAppFrame
 		MakeNewTab("Search panel", new SearchPanel(this));
 	}
 	
+	/**
+	 * Return Id of current user
+	 * @return
+	 */
 	public String getUserId() {
 		if(user!=null) return user.getId();
 		return null;
 	}
 	
+	/**
+	 * Gets all film data from database and make instance of Film in different thread
+	 * If data not exists, invoke WebService and save data in database
+	 * for repeated use
+	 * 
+	 * @param tconst
+	 * @return
+	 */
 	Film MakeNewFilm(String tconst)
 	{
 		HashMap<String, String> set=null;
@@ -58,6 +78,9 @@ public class ApplicationWindow extends SearchAppFrame
 		return null;
 	}
 	
+	/**
+	 * Launches the authorization procedure
+	 */
 	void logIn() {
 		String s = loginDialog(this);
 		if(s!=null) {
@@ -70,10 +93,16 @@ public class ApplicationWindow extends SearchAppFrame
 		}
 	}
 
+	/**
+	 * Open form for user registration
+	 */
 	void register() {
-		MakeNewTab("User panel", user);
+		if(user.level==0) MakeNewTab("User panel", user);
 	}
 
+	/**
+	 * Launches the unauthorization procedure
+	 */
 	void logOut() {
 		if(user.getId().length()>0)dbService.logOut(user.getId());
 		user=new User(this);
@@ -82,12 +111,21 @@ public class ApplicationWindow extends SearchAppFrame
 		levelState();
 	}
 
+	/**
+	 * Check access level of current user and repaint window according to it
+	 */
 	private void levelState() {
 		menu.state(user.level);
 		revalidate();
 		repaint();
 	}
 
+	/**
+	 * Show form with all data of current user
+	 * boolean parameter switches to edit mode
+	 * 
+	 * @param i
+	 */
 	void showUserData(boolean i) {
 		RegisteredUser tmp = (RegisteredUser) user;
 		tmp.detailsSetEditable(i);
@@ -97,7 +135,7 @@ public class ApplicationWindow extends SearchAppFrame
 	}
 	
 	/**
-	 * 
+	 * Show current no submited order if exists
 	 */
 	void showCurrentOrder() {
 		if(order==null || order.orderedInStock.size()==0) {
@@ -109,16 +147,23 @@ public class ApplicationWindow extends SearchAppFrame
 	}
 	
 	/**
-	 * Returns inner inventory code of film
+	 * Invoke DbServise to book film by IMDB code
+	 * Returns specific inner inventory code of booked film
 	 * 
 	 * @param tconst imdb code
-	 * 
 	 * @return invconst inner inventory code
 	 */
 	public String getInventory(String tconst) {
 		return dbService.getInventory(tconst);
 	}
 	
+	/**
+	 * Open login dialog and invoke FileSystemService
+	 *  to check inputed data
+	 * 
+	 * @param frame
+	 * @return registered user email if done
+	 */
 	static String loginDialog(JFrame frame) {
 		
 	    JPanel panel = new JPanel(new BorderLayout(5, 5));
@@ -141,6 +186,13 @@ public class ApplicationWindow extends SearchAppFrame
 	    return null;
 	}
 	
+	/**
+	 * Gets all update data of registered user 
+	 * and completes update procedure
+	 * 
+	 * @param set
+	 * @return true if done
+	 */
 	boolean updateUserRegistration(HashMap<String, String> set) {
 		RegisteredUser u = new RegisteredUser(this,set);
 		 if(dbService.updateUser(u)) {
@@ -152,6 +204,14 @@ public class ApplicationWindow extends SearchAppFrame
 		 }else return false;
 	}
 	
+	/**
+	 * Gets all data from registration form
+	 * and completes registration procedure
+	 * includes store password
+	 * 
+	 * @param set
+	 * @return true if done
+	 */
 	void newUserRegistration(HashMap<String, String> set) {
 		Thread thread = new Thread(){
 		    public void run(){
@@ -170,16 +230,36 @@ public class ApplicationWindow extends SearchAppFrame
 		 }
 	}
 
+	/**
+	 * Invoke FileSystemService to change user password
+	 * 
+	 * @param s
+	 * @return
+	 */
 	boolean updatePassword(String[] s) {
 		if(FileSystemService.updatePasswordInFile(s)) return true;
 		return false;
 	}
 	
+	/**
+	 *  Invoke FileSystemService to change user email in password storage
+	 *  
+	 * @param oldEmail
+	 * @param newEmail
+	 * @param pass
+	 * @return
+	 */
 	boolean changeEmailInPassword(String oldEmail, String newEmail, String pass) {
 		if(FileSystemService.updateEmailInFile(oldEmail, newEmail, pass))return true;
 		else return false;				
 	}
 	
+	/**
+	 * Invoke infoMsgByPlace function for temporary show given message on top of window
+	 * 
+	 * @param msg
+	 * @param seconds
+	 */
 	void infoMsgTop(String msg, int seconds) {
 		StringBuilder s = new StringBuilder();
 		for(int i=0;i<50-msg.length()/2;i++) {
@@ -189,42 +269,88 @@ public class ApplicationWindow extends SearchAppFrame
 		infoMsgByPlace(menu.infoLabel, s.toString(), seconds);
 	}
 	
+	/**
+	 * Launches exit procedure
+	 * 
+	 * @return
+	 */
 	int exit() {
 		dbService.logOut(user.getId());
 		System.exit(0);
 		return 0;
 	}
 
+	/**
+	 * Book new order number in database and return its
+	 * @return
+	 */
 	int getOrderNumber() {
 		return dbService.newOrderNumber(user.getId());
 	}
 
+	/**
+	 * Submit order in database
+	 * 
+	 * @param orderNumber
+	 * @param orderedInStock
+	 * @return
+	 */
 	boolean submitOrder(int orderNumber, HashMap<String, Film> orderedInStock) {
 		if(dbService.submitOrder(orderNumber, orderedInStock))return true;
 		return false;
 	}
 
+	/**
+	 * Invoke pay procedure(not implements in this project) 
+	 * and set order as payed if done.
+	 * 
+	 * @param orderNumber
+	 * @return
+	 */
 	boolean pay(int orderNumber) {
 		if(PaymentService.pay() && dbService.setPayed(orderNumber)) return true;
 		return false;
 	}
 
+	/**
+	 * Make new instance of OrderPanel
+	 */
 	void newOrder() {
 		removeTabByName("Order");
 		order = new OrderPanel(this);
 	}
 	
+	/**
+	 * Cancellation of booking film in database
+	 * 
+	 * @param inv
+	 * @return
+	 */
 	boolean removeOrderFilmFromDB(String inv) {
 		if(dbService.removeOrderFilmFromDB(inv, user.getId())) return true;
 		return false;
 	}
 	
+	/**
+	 * Booking film to user in database
+	 * @param film
+	 * @return
+	 */
 	boolean addFilmToOrder(Film film) {
 		String inv = dbService.orderFilm(film, user.getId());
 		if(inv!= null && order.addFilmToOrder(film, inv)) return true;
 		return false;
 	}
 
+	/**
+	 * Invoke getAdminSearch in DbSerice for extended search data
+	 * in admin mode
+	 * 
+	 * @param mode
+	 * @param type
+	 * @param input
+	 * @return
+	 */
 	public HashMap<String,String> getAdminSearch(String mode, String type, String input) {
 		if(user.level<2) return null;
 		
@@ -232,30 +358,54 @@ public class ApplicationWindow extends SearchAppFrame
 		return result;
 	}
 
+	/**
+	 * Open admin search panel
+	 */
 	public void showAdminPanel() {
 		removeTabByName("Admin panel");
 		MakeNewTab("Admin panel", new AdminPanel(this));
 	}
-	
-	public static void main(String[] args)
-	{
-		ApplicationWindow newApp = new ApplicationWindow();
-	}
 
+	/**
+	 * Delete film from base(admin mode)
+	 * 
+	 * @param tconst
+	 */
 	public void deleteFilm(String tconst) {
+		if(user.level<2) return;
 		dbService.deleteFilm(tconst);
 		
 	}
 
+	/**
+	 * Delete user from base(admin mode)
+	 * 
+	 * @param id
+	 */
 	public void deleteUser(String id) {
+		if(user.level<2) return;
 		dbService.deleteUser(id);
 	}
 
+	/**
+	 * Mark all ordered films as returned(admin mode)
+	 * 
+	 * @param orderNumber
+	 */
 	public void closeOrder(int orderNumber) {
+		if(user.level<2) return;
 		dbService.closeOrder(orderNumber);
 	}
 
+	/**
+	 * Get all ordered films from base by order number
+	 * (admin mode)
+	 * 
+	 * @param orderNumber
+	 * @return
+	 */
 	public HashMap<String, Film> getOrderedFilms(int orderNumber) {
+		if(user.level<2) return null;
 		Film f = null;
 		Vector<String> filmList = dbService.getOrderedFilms(orderNumber);
 		HashMap<String,Film> orderedInStock = new HashMap<String,Film>();
@@ -267,11 +417,36 @@ public class ApplicationWindow extends SearchAppFrame
 		return orderedInStock;
 	}
 
+	/**
+	 * Get pay status of cpecific order
+	 * (admin mode)
+	 * 
+	 * @param orderNumber
+	 * @return
+	 */
 	public boolean getPayed(int orderNumber) {
+		if(user.level<2) return false;
 		return dbService.getPayed(orderNumber);
 	}
 
+	/**
+	 * Get all registered user details by user ID
+	 * (admin mode)
+	 * 
+	 * @param id
+	 * @return
+	 */
 	public HashMap<String, String> getUserById(String id) {
+		if(user.level<2) return null;
 		return dbService.getUserById(id);
+	}
+	
+	/**
+	 * Application entry function
+	 * @param args
+	 */
+	public static void main(String[] args)
+	{
+		ApplicationWindow newApp = new ApplicationWindow();
 	}
 }
